Revert "Unify data types for the statistics data"
This reverts commit c6566ea7fa07c2334a5e4abfc8c1275dd41b4137.

+++ main/tasks/statistics_task.h
@@ -7,17 +7,17 @@ typedef struct StatisticsData * StatisticsNextNodePtr;
 struct StatisticsData
 {
     int64_t timestamp;
-    float hashrate;
+    double hashrate;
     float chipTemperature;
     float vrTemperature;
     float power;
     float voltage;
     float current;
-    float coreVoltageActual;
-    float fanSpeed;
-    float fanRPM;
-    float wifiRSSI;
-    float freeHeap;
+    int16_t coreVoltageActual;
+    uint16_t fanSpeed;
+    uint16_t fanRPM;
+    int8_t wifiRSSI;
+    uint32_t freeHeap;
 
     StatisticsNextNodePtr next;
 };
